feat(column): add fractional position reordering and fix CORS PATCH support

+++ src/main/java/com/billyow/app/boardang/boardColumn/DTO/BoardColumnResponse.java
@@ -7,6 +7,6 @@ import java.util.List;
 public record BoardColumnResponse(
         Long id,
         String title,
-        Integer position,
+        Double position,
         List<TaskResponse> tasks
 ){}
