+++ src/main/java/com/billyow/app/boardang/boardColumn/DTO/BoardColumnUpdateRequest.java
@@ -2,6 +2,6 @@ package com.billyow.app.boardang.boardColumn.DTO;
 
 public record BoardColumnUpdateRequest(
         String title,
-        Integer position
+        Double position
         ) {
 }
