+++ src/main/java/com/billyow/app/boardang/boardColumn/repository/IBoardColumnRepository.java
@@ -13,7 +13,7 @@ public interface IBoardColumnRepository extends JpaRepository<BoardColumn, Long>
     SELECT MAX(c.position) FROM BoardColumn c
     WHERE c.board.id = :boardId
 """)
-     Integer getMaxPositionByBoardId(@Param("boardId")Long boardId);
+     Double getMaxPositionByBoardId(@Param("boardId")Long boardId);
 
      Integer countByBoard_Id(Long boardId);
